server/requirements.txt
fastapi==0.115.8
uvicorn==0.30.0
pydantic==2.6.1
python-dotenv==1.0.0
langchain-google-genai==0.0.5
langchain-core==0.1.17
langchain-chroma==0.0.16
langchain-community==0.0.13
scikit-learn==1.3.2
numpy==1.26.3
scholarly==1.8.1
duckduckgo-search==3.9.9
python-multipart==0.0.9
